# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinZ LMS API",
        "description": "API documentation for FinZ LMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/manual-reviews": {
            "get": {
                "tags": [
                    "ManualReview"
                ],
                "summary": "index ManualReview",
                "operationId": "eaa724c488e77470adc6600a3711a4b4",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/manual-reviews/{id}": {
            "get": {
                "tags": [
                    "ManualReview"
                ],
                "summary": "show ManualReview",
                "operationId": "15bf2365fb71ba497071d99e7e951635",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/manual-reviews/{id}/decide": {
            "post": {
                "tags": [
                    "ManualReview"
                ],
                "summary": "decide ManualReview",
                "operationId": "b94d1512c5bf32619ac2a3cc1d5caf3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/manual-reviews/scorecard/{reviewer_id}": {
            "get": {
                "tags": [
                    "ManualReview"
                ],
                "summary": "scorecard ManualReview",
                "operationId": "c244d76479e60af8bb5b21fbac185cd0",
                "parameters": [
                    {
                        "name": "reviewer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/risk-rules": {
            "get": {
                "tags": [
                    "RiskRule"
                ],
                "summary": "index RiskRule",
                "operationId": "1bd779d89623bde4647b8c59da12f768",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RiskRule"
                ],
                "summary": "store RiskRule",
                "operationId": "a5e275bd9b309ae03c7a27eb5957c7f2",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/risk-rules/{id}": {
            "put": {
                "tags": [
                    "RiskRule"
                ],
                "summary": "update RiskRule",
                "operationId": "3b3612af7c6d9b8c05c72fa8ff87e4d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/risk-rules/simulate": {
            "post": {
                "tags": [
                    "RiskRule"
                ],
                "summary": "simulate RiskRule",
                "operationId": "2db80edcbabd4627f12574e42af9ce16",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}